Another round of Java generics addition, and various other minor changes.
git-svn-id: https://jcae.svn.sourceforge.net/svnroot/jcae/trunk@1605 9c4724a3-da1e-0410-bc68-88f18da35cd4

+++ amibe/src/org/jcae/mesh/MeshValidSoup.java
@@ -64,7 +64,7 @@ public class MeshValidSoup
 			a1 *= 3.0 / Math.PI;
 			if (a1 < amin[group])
 				amin[group] = a1;
-			int index = (int) (a1 * (double) split);
+			int index = (int) (a1 * split);
 			nr[group][index]++;
 		}
 		public void printStats()
